+++ src/CMakeLists.txt
@@ -57,8 +57,6 @@ set(Mapper_Common_SRCS
  util_task_dialog.cpp
  util_translation.cpp
  mapper_resource.cpp
- undo.cpp
- undo_manager.cpp
  matrix.cpp
  transformation.cpp
 
@@ -66,8 +64,10 @@ set(Mapper_Common_SRCS
 
  touch_cursor.cpp
 
- object_undo.cpp
- map_part_undo.cpp
+ undo/undo.cpp
+ undo/undo_manager.cpp
+ undo/object_undo.cpp
+ undo/map_part_undo.cpp
 
  color_dock_widget.cpp
 
@@ -209,7 +209,6 @@ set(Mapper_Common_MOC_INPUT
  sensors/gps_temporary_markers.h
  sensors/gps_track_recorder.h
  core/map.h
- object_undo.h
  settings.h
  core/symbols/area_symbol.h
  core/symbols/combined_symbol.h
@@ -248,8 +247,10 @@ set(Mapper_Common_MOC_INPUT
  tools/rotate_tool.h
  tools/rotate_pattern_tool.h
  tools/scale_tool.h
- undo_manager.h
  util_task_dialog.h
+
+ undo/undo_manager.h
+ undo/object_undo.h
  
  core/autosave_p.h
  core/georeferencing.h
@@ -344,12 +345,12 @@ set(Mapper_Common_HEADERS
   util/scoped_signals_blocker.h
 
   core/map_part.h
-  map_part_undo.h
   core/objects/object_operations.h
   core/renderables/renderable.h
   core/renderables/renderable_implementation.h
   core/symbols/symbol.h
-  undo.h
+  undo/map_part_undo.h
+  undo/undo.h
   util_gui.h
 )
