+++ src/gui/map/map_editor.cpp
@@ -70,8 +70,8 @@
 #include "map_dialog_rotate.h"
 #include "map_dialog_scale.h"
 #include "map_editor_activity.h"
-#include "map_part_undo.h"
-#include "object_undo.h"
+#include "undo/map_part_undo.h"
+#include "undo/object_undo.h"
 #include "map_widget.h"
 #include "sensors/gps_display.h"
 #include "sensors/gps_temporary_markers.h"
@@ -115,7 +115,7 @@
 #include "tools/rotate_tool.h"
 #include "tools/rotate_pattern_tool.h"
 #include "tools/scale_tool.h"
-#include "undo_manager.h"
+#include "undo/undo_manager.h"
 #include "util.h"
 #include "util/backports.h"
 #include "util/scoped_signals_blocker.h"
