+++ src/tools/boolean_tool.cpp
@@ -29,7 +29,7 @@
 #include "core/map.h"
 #include "core/symbols/symbol.h"
 #include "core/objects/object.h"
-#include "object_undo.h"
+#include "undo/object_undo.h"
 #include "util.h"
