+++ src/tools/cut_hole_tool.cpp
@@ -26,7 +26,7 @@
 #include <QMessageBox>
 
 #include "core/map.h"
-#include "object_undo.h"
+#include "undo/object_undo.h"
 #include "gui/map/map_widget.h"
 #include "core/objects/object.h"
 #include "core/symbols/symbol.h"
