+++ src/tools/cut_tool.cpp
@@ -28,7 +28,7 @@
 
 #include "core/map.h"
 #include "gui/map/map_widget.h"
-#include "object_undo.h"
+#include "undo/object_undo.h"
 #include "core/renderables/renderable.h"
 #include "settings.h"
 #include "core/symbols/symbol.h"
