+++ src/tools/draw_line_and_area_tool.cpp
@@ -25,7 +25,7 @@
 
 #include "core/map.h"
 #include "gui/map/map_editor.h"
-#include "object_undo.h"
+#include "undo/object_undo.h"
 #include "gui/map/map_widget.h"
 #include "core/objects/object.h"
 #include "core/renderables/renderable.h"
