+++ src/tools/edit_line_tool.cpp
@@ -27,7 +27,7 @@
 #include <QKeyEvent>
 
 #include "core/map.h"
-#include "object_undo.h"
+#include "undo/object_undo.h"
 #include "gui/map/map_widget.h"
 #include "core/objects/object.h"
 #include "core/objects/text_object.h"
