+++ test/symbol_set_t.cpp
@@ -27,7 +27,7 @@
 #include "core/symbols/area_symbol.h"
 #include "../src/settings.h"
 #include "../src/templates/template.h"
-#include "../src/undo_manager.h"
+#include "undo/undo_manager.h"
 
 
 /**
